requirements.txt
requests==2.31.0
tenacity==8.2.3
pandas==2.1.4
pyarrow==14.0.2
pyspark==3.5.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
boto3==1.34.0
python-dotenv==1.0.0
apache-airflow==2.8.1
